feat: allow overriding of persistence properties
Prior to this change, the only way to change the persistence settings
was to edit the `persistence.xml`. In case of a pre-build .jar, the
`persistence.xml` has to be contained in the .jar with the persistence
module [1], and can not be overridden. However, the current repo
hard-codes the DB connection details in the persistence.xml, which
makes it impossible to override the DB connection details for a fully
built .jar without changing that .jar file.

This change enable the overriding of the properties used for the JPA
`EntityManagerFactory` with system properties, so that DB connection
details can be altered without changing the .jar file.

[1] https://openjpa.apache.org/builds/1.0.2/apache-openjpa-1.0.2/docs/manual/ref_guide_conf_specify.html